refactor(server): 优化server结构
- dispatch_console_command 从 methods/server/console.rs 迁入 service/console.rs，
  与服务层职责对齐；methods/server/ 现仅包含纯 RPC 方法定义
- 删除 methods/server/console.rs，统一为 console_command.rs 一个方法文件
- 更新 service/mod.rs 暴露 dispatch_console_command，路径变更为
  sealantern_server::rpc::service::dispatch_console_command
- 更新 src-tauri 中导入路径适配上述变更
- 统一 PERMISSION_SERVER_CONSOLE_SEND 导出路径至 methods:: 层级
- console_command.rs 新增完整的接口文档注释（HTTP 方法、路径、请求体、响应体）
