+++ server/src/rpc/methods/server/mod.rs
@@ -1,8 +1,5 @@
 //! 服务器实例管理方法。
 
-mod console;
 mod console_command;
 
-pub use super::permissions::PERMISSION_SERVER_CONSOLE_SEND;
-pub use console::dispatch_console_command;
 pub use console_command::{ConsoleCommandRequest, SendConsoleCommand};
